PlantClima.jl

Index

PlantClima.simple_Ta_RH

PlantClima.simple_Ta_RH_Q

API documentation

Un-exported

Private functions, types or constants from PlantClima. These are not exported, so you need to prefix the function name with PlantClima. to access them:

PlantClima.simple_Ta_RH — Method

simpleTaRH(; lat::Float64, DOY::Float64, h::Float64, Tmin::Vector{Float64}, Tmax::Vector{Float64}, ea::Vector{Float64}, p::Float64 = 1.5, tc::Float64 = 4.0)

Arguments:

- `lat`: latitide, []
- `DOY`: day of the year, [d]
- `h`: hour of the day, [h]
- `Tmin`: vector of daily minimum temperature for the previous day, current day, and next day, [°C]
- `Tmax`: vector of daily maximum temperature for the previous day, current day, and next day, [°C]
- `ea`: vector of daily average vapor pressure for the previous day, current day, and next day, [kPa]

Citation: Goudriaan, J., & van Laar, H. H. (1994). Modelling potential crop growth processes. Textbook with exercises. Springer Science & Business Media.

PlantClima.simple_Ta_RH_Q — Method

simple_Ta_RH(; lat::W, DOY::W, h::W, 
               Tmin::Vector{T}, Tmax::Vector{T}, ea::Vector{P}, 
               p::W = 1.5, tc::W = 4.0) where {T <: Quantity, P <: Quantity, W <: Real}

Arguments:

- `lat`: latitide, []
- `DOY`: day of the year, [d]
- `h`: hour of the day, [h]
- `Tmin`: vector of daily minimum temperature for the previous day, current day, and next day, [K]
- `Tmax`: vector of daily maximum temperature for the previous day, current day, and next day, [K]
- `ea`: vector of daily average vapor pressure for the previous day, current day, and next day, [Pa]